+++ src/lib/common/Utils.ts
@@ -1156,6 +1156,22 @@ function getObjectFromVariable(variable){
 
 type JsonObject = { [key: string]: any };
 
+function getAllValuesFromPropInJSONByPropertyName(obj: JsonObject, propertyName: string, maxDepth: number = 10, cond : (key, value, tabToSaveObjFound) => void = null): any[] | any {
+    let valuesObjs = findAllPropInJSONByPropertyName(obj, propertyName, maxDepth, cond );
+    let values= undefined;
+    if (Array.isArray(valuesObjs) === true) {
+        if (valuesObjs.length > 0) {
+            values = [];
+            for (let i = 0; i < valuesObjs.length; i++) {
+                values.push(getTextFromJSONProperty(valuesObjs[i]));
+            }
+        }
+    } else if (valuesObjs !== undefined) {
+        values = getTextFromJSONProperty(valuesObjs);
+    }
+    return values;
+}
+
 function findAllPropInJSONByPropertyName(obj: JsonObject, propertyName: string, maxDepth: number = 10, cond : (key, value, tabToSaveObjFound) => void = null): any[] | any {
     let results: any[] = [];
 
@@ -1526,6 +1542,7 @@ export let objToExport = {
     getAlternateMessageFromJSONObj,
     getValueFromVariable,
     getObjectFromVariable,
+    getAllValuesFromPropInJSONByPropertyName,
     findAllPropInJSONByPropertyName,
     findAllPropInJSONByPropertyNameByXmlNS,
     callerName,
@@ -1600,6 +1617,7 @@ export {
     getAlternateMessageFromJSONObj,
     getValueFromVariable,
     getObjectFromVariable,
+    getAllValuesFromPropInJSONByPropertyName,
     findAllPropInJSONByPropertyName,
     findAllPropInJSONByPropertyNameByXmlNS,
     callerName,
@@ -1673,6 +1691,7 @@ export default {
     getAlternateMessageFromJSONObj,
     getValueFromVariable,
     getObjectFromVariable,
+    getAllValuesFromPropInJSONByPropertyName,
     findAllPropInJSONByPropertyName,
     findAllPropInJSONByPropertyNameByXmlNS,
     callerName,